fix(core): give IntPtr the defined native-width wrap .NET relies on
CCF-004 class A, the family's second implemented member. SR-AUD-025 is now
remediated; the findings index moves to 23 remediated / 341 confirmed of 364.

IntPtr::Add evaluated `pointer.value + offset` in intptr_t and Subtract the
mirror, so both native boundary cases were signed-overflow UB
(build-probe/1829_ccf004_survey.log cases 3 and 4). .NET exposes these as
`pointer + offset` / `pointer - offset` for nint under ordinary unchecked C#
arithmetic, whose modulo-native-width wrap is defined. Both now compute in
uintptr_t.

Site enumeration used the recovering build first, per the amendment #1830 added
as docs/DefinedArithmeticBoundaryPlan.md section 12, and confirmed exactly two
sites. The friend operator+ / operator- forms forward to the named methods rather
than duplicating the arithmetic, and are pinned independently so a refactor that
stops forwarding cannot silently reintroduce the overflow.

One subtlety is recorded because getting it wrong would have been a regression
rather than a neutral difference: the offset needs a two-step cast, intcs ->
intptr_t -> uintptr_t. intcs is 32-bit and uintptr_t is 64-bit on LP64, so a
negative offset must be sign-extended before widening; casting straight to
uintptr_t zero-extends it and turns Add(p, -1) into an addition of 4294967295.
Pinned by dedicated negative-offset tests in both methods and both operator forms.

No observable change, proven: MaxValue + 1 == MinValue and MinValue - 1 ==
MaxValue are the values measured before the repair, and
build-probe/1832_postfix.log shows the UB absent with the values byte-identical.

9 permanent regressions. SharpRuntimeTests_Core_Base 5009/5009, clean under
ASan + UBSan + LSan with 0 reports. Repository gate 14,113 tests across 37
executables, 0 warnings, 0 errors; local_ci_check.sh build passed. No public
signature, virtual, vtable, object layout or mangled symbol changed.

Ticket: #1832 (REMED-CORE-INTPTR-DEFINED-WRAP, P1). Remediates SR-AUD-025.

+++ modules/core/include/System/IntPtr.hpp
@@ -100,18 +100,42 @@ namespace System
          * @brief Adds an offset to the value of a pointer.
          *
          * C++ counterpart of .NET IntPtr.Add(nint, int) / operator+(nint, int).
+         *
+         * The addition is performed in `uintptr_t` and converted back (ticket #1832,
+         * CCF-004 class A). .NET exposes this as `pointer + offset` for `nint` under
+         * ordinary **unchecked** C# arithmetic, whose modulo-native-width wrap is
+         * *defined*; the signed C++ form was not. `Add(MaxValue, 1)` was UBSan-confirmed
+         * undefined behaviour at this line (`build-probe/1829_ccf004_survey.log` case 3,
+         * `signed integer overflow: 9223372036854775807 + 1`). The value is unchanged —
+         * `MaxValue + 1 == MinValue` before and after, pinned by a test.
          */
         [[nodiscard]] static IntPtr Add(const IntPtr& pointer, intcs offset) {
-            return IntPtr(static_cast<intptr_t>(pointer.value + offset));
+            return IntPtr(static_cast<intptr_t>(static_cast<uintptr_t>(pointer.value)
+                                                + static_cast<uintptr_t>(
+                                                      static_cast<intptr_t>(offset))));
         }
 
         /**
          * @brief Subtracts an offset from the value of a pointer.
          *
          * C++ counterpart of .NET IntPtr.Subtract(nint, int) / operator-(nint, int).
+         *
+         * The subtraction is performed in `uintptr_t` and converted back, for the same
+         * reason as Add above (ticket #1832, CCF-004 class A). `Subtract(MinValue, 1)` was
+         * UBSan-confirmed undefined behaviour at this line (case 4,
+         * `signed integer overflow: -9223372036854775808 - 1`). The value is unchanged —
+         * `MinValue - 1 == MaxValue` before and after.
+         *
+         * Note the two-step cast on @p offset in both methods: `intcs` is 32-bit and
+         * `uintptr_t` is 64-bit on LP64, so a **negative** offset must be sign-extended to
+         * `intptr_t` *first*. Converting straight to `uintptr_t` would zero-extend it and
+         * turn `Add(p, -1)` into an addition of 4294967295 — a wrong answer, not merely a
+         * differently-arrived-at one. Pinned by the negative-offset tests.
          */
         [[nodiscard]] static IntPtr Subtract(const IntPtr& pointer, intcs offset) {
-            return IntPtr(static_cast<intptr_t>(pointer.value - offset));
+            return IntPtr(static_cast<intptr_t>(static_cast<uintptr_t>(pointer.value)
+                                                - static_cast<uintptr_t>(
+                                                      static_cast<intptr_t>(offset))));
         }
 
         /** @brief Adds an offset to the value of a pointer. C++ counterpart of .NET IntPtr operator+(nint, int). */
